chore: 상품 등록 페이지 Main 리팩토링 (#31)
